# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TurboStream API for real-time data streaming and feed management with AI-powered analysis",
        "title": "TurboStream API",
        "termsOfService": "https://turboline.ai/terms",
        "contact": {
            "name": "API Support",
            "url": "https://turboline.ai/docs/api",
            "email": "dev@turboline.ai"
        },
        "license": {
            "name": "MPL-2.0",
            "url": "https://github.com/turboline-ai/turbostream?tab=MPL-2.0-1-ov-file"
        },
        "version": "1.0"
    },
    "host": "localhost:7210",
    "basePath": "/",
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "description": "Register a new user account with email, password, and name",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Authentication"],
                "summary": "Register new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "description": "Login with email and password. Returns JWT token for authentication. Supports 2FA with TOTP token.",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["Authentication"],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/AuthResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "security": [{"BearerAuth": []}],
                "description": "Get current authenticated user profile with token usage information",
                "produces": ["application/json"],
                "tags": ["Authentication"],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RegisterRequest": {
            "type": "object",
            "required": ["email", "password", "name"],
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "example": "securePassword123"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                }
            }
        },
        "LoginRequest": {
            "type": "object",
            "required": ["email", "password"],
            "properties": {
                "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "format": "password",
                    "example": "securePassword123"
                },
                "totpToken": {
                    "type": "string",
                    "description": "TOTP code if 2FA is enabled",
                    "example": "123456"
                }
            }
        },
        "AuthResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "token": {
                    "type": "string",
                    "description": "JWT authentication token",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
        "UserProfileResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
        "User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "507f1f77bcf86cd799439011"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "createdAt": {
                    "type": "string",
                    "format": "date-time"
                },
                "tokenUsage": {
                    "$ref": "#/definitions/TokenUsage"
                },
                "twoFactorEnabled": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "TokenUsage": {
            "type": "object",
            "properties": {
                "currentMonth": {
                    "type": "string",
                    "example": "2025-01"
                },
                "tokensUsed": {
                    "type": "integer",
                    "example": 1500
                },
                "limit": {
                    "type": "integer",
                    "example": 1000000
                },
                "lastResetDate": {
                    "type": "string",
                    "format": "date-time"
                },
                "overdraftAllowed": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Error message describing what went wrong"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type 'Bearer' followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}